ENGINES: Extract more base objects for KotOR games
Base objects extracted: situated, placeable, door, trigger, waypoint,
sound and pathfinding classes.
