+++ src/engines/kotor/area.cpp
@@ -45,18 +45,19 @@
 
 #include "src/engines/kotorbase/room.h"
 #include "src/engines/kotorbase/creature.h"
+#include "src/engines/kotorbase/placeable.h"
+#include "src/engines/kotorbase/door.h"
+#include "src/engines/kotorbase/waypoint.h"
+#include "src/engines/kotorbase/trigger.h"
+#include "src/engines/kotorbase/sound.h"
+
+#include "src/engines/kotorbase/path/pathfinding.h"
+#include "src/engines/kotorbase/path/objectwalkmesh.h"
+#include "src/engines/kotorbase/path/doorwalkmesh.h"
 
 #include "src/engines/kotor/area.h"
 #include "src/engines/kotor/module.h"
-#include "src/engines/kotor/waypoint.h"
-#include "src/engines/kotor/placeable.h"
-#include "src/engines/kotor/pathfinding.h"
-#include "src/engines/kotor/objectwalkmesh.h"
-#include "src/engines/kotor/door.h"
-#include "src/engines/kotor/doorwalkmesh.h"
-#include "src/engines/kotor/trigger.h"
 #include "src/engines/kotor/actionexecutor.h"
-#include "src/engines/kotor/sound.h"
 
 namespace Engines {
