--- src/engines/kotor/pathfinding.h
@@ -1,67 +1,0 @@
-/* xoreos - A reimplementation of BioWare's Aurora engine
- *
- * xoreos is the legal property of its developers, whose names
- * can be found in the AUTHORS file distributed with this source
- * distribution.
- *
- * xoreos is free software; you can redistribute it and/or
- * modify it under the terms of the GNU General Public License
- * as published by the Free Software Foundation; either version 3
- * of the License, or (at your option) any later version.
- *
- * xoreos is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with xoreos. If not, see <http://www.gnu.org/licenses/>.
- */
-
-/** @file
- *  Pathfinding class for KotOR.
- */
-
-#ifndef ENGINES_KOTOR_PATHFINDING_H
-#define ENGINES_KOTOR_PATHFINDING_H
-
-#include "glm/mat4x4.hpp"
-
-#include "src/aurora/types.h"
-
-#include "src/engines/aurora/pathfinding.h"
-
-#include "src/engines/kotor/walkmeshloader.h"
-
-namespace Engines {
-
-namespace KotOR {
-
-class WalkmeshLoader;
-class Room;
-
-class Pathfinding : public Engines::Pathfinding {
-public:
-	Pathfinding(const std::vector<bool> &walkableProp);
-	~Pathfinding();
-
-	void addRoom(Room *room);
-	void connectRooms();
-
-	Room *getRoomAt(float x, float y) const;
-
-protected:
-	WalkmeshLoader _walkmeshLoader;
-	std::vector<std::map<uint32, uint32> > _adjRooms;
-	std::vector<uint32> _startFace;
-
-private:
-	uint32 getFaceFromEdge(uint32 edge, uint32 room) const;
-	std::vector<Room *> _rooms;
-};
-
-} // End of namespace KotOR
-
-} // End of namespace Engines
-
-#endif // ENGINES_KOTOR_PATHFINDING_H
