+++ src/engines/kotor/rules.mk
@@ -29,22 +29,12 @@ src_engines_kotor_libkotor_la_SOURCES += \
     src/engines/kotor/modelloader.h \
     src/engines/kotor/version.h \
     src/engines/kotor/game.h \
-    src/engines/kotor/situated.h \
-    src/engines/kotor/waypoint.h \
-    src/engines/kotor/placeable.h \
-    src/engines/kotor/door.h \
-    src/engines/kotor/sound.h \
     src/engines/kotor/area.h \
     src/engines/kotor/objectcontainer.h \
     src/engines/kotor/module.h \
     src/engines/kotor/savedgame.h \
-    src/engines/kotor/trigger.h \
     src/engines/kotor/actionexecutor.h \
     src/engines/kotor/partyconfig.h \
-    src/engines/kotor/pathfinding.h \
-    src/engines/kotor/walkmeshloader.h \
-    src/engines/kotor/objectwalkmesh.h \
-    src/engines/kotor/doorwalkmesh.h \
     $(EMPTY)
 
 src_engines_kotor_libkotor_la_SOURCES += \
@@ -54,22 +44,12 @@ src_engines_kotor_libkotor_la_SOURCES += \
     src/engines/kotor/modelloader.cpp \
     src/engines/kotor/version.cpp \
     src/engines/kotor/game.cpp \
-    src/engines/kotor/waypoint.cpp \
-    src/engines/kotor/situated.cpp \
-    src/engines/kotor/placeable.cpp \
-    src/engines/kotor/door.cpp \
-    src/engines/kotor/sound.cpp \
     src/engines/kotor/area.cpp \
     src/engines/kotor/objectcontainer.cpp \
     src/engines/kotor/module.cpp \
     src/engines/kotor/savedgame.cpp \
-    src/engines/kotor/trigger.cpp \
     src/engines/kotor/actionexecutor.cpp \
     src/engines/kotor/partyconfig.cpp \
-    src/engines/kotor/pathfinding.cpp \
-    src/engines/kotor/walkmeshloader.cpp \
-    src/engines/kotor/objectwalkmesh.cpp \
-    src/engines/kotor/doorwalkmesh.cpp \
     $(EMPTY)
 
 include src/engines/kotor/script/rules.mk
