+++ src/engines/kotor/script/functions_action.cpp
@@ -29,11 +29,11 @@
 #include "src/aurora/nwscript/functioncontext.h"
 
 #include "src/engines/kotorbase/types.h"
+#include "src/engines/kotorbase/door.h"
 
 #include "src/engines/kotor/game.h"
 #include "src/engines/kotor/module.h"
 #include "src/engines/kotor/objectcontainer.h"
-#include "src/engines/kotor/door.h"
 
 #include "src/engines/kotor/script/functions.h"
