+++ src/engines/kotor/script/functions_object.cpp
@@ -31,11 +31,11 @@
 
 #include "src/engines/kotorbase/types.h"
 #include "src/engines/kotorbase/object.h"
+#include "src/engines/kotorbase/placeable.h"
 
 #include "src/engines/kotor/game.h"
 #include "src/engines/kotor/module.h"
 #include "src/engines/kotor/objectcontainer.h"
-#include "src/engines/kotor/placeable.h"
 #include "src/engines/kotor/area.h"
 
 #include "src/engines/kotor/script/functions.h"
