+++ src/engines/kotor/script/functions_situated.cpp
@@ -25,9 +25,9 @@
 #include "src/aurora/nwscript/functioncontext.h"
 
 #include "src/engines/kotorbase/object.h"
+#include "src/engines/kotorbase/situated.h"
 
 #include "src/engines/kotor/objectcontainer.h"
-#include "src/engines/kotor/situated.h"
 
 #include "src/engines/kotor/script/functions.h"
