+++ src/engines/kotor2/rules.mk
@@ -28,18 +28,10 @@ src_engines_kotor2_libkotor2_la_SOURCES += \
     src/engines/kotor2/console.h \
     src/engines/kotor2/modelloader.h \
     src/engines/kotor2/game.h \
-    src/engines/kotor2/waypoint.h \
-    src/engines/kotor2/situated.h \
-    src/engines/kotor2/placeable.h \
-    src/engines/kotor2/door.h \
     src/engines/kotor2/creature.h \
     src/engines/kotor2/area.h \
     src/engines/kotor2/objectcontainer.h \
     src/engines/kotor2/module.h \
-    src/engines/kotor2/trigger.h \
-    src/engines/kotor2/pathfinding.h \
-    src/engines/kotor2/objectwalkmesh.h \
-    src/engines/kotor2/doorwalkmesh.h \
     $(EMPTY)
 
 src_engines_kotor2_libkotor2_la_SOURCES += \
@@ -48,18 +40,10 @@ src_engines_kotor2_libkotor2_la_SOURCES += \
     src/engines/kotor2/console.cpp \
     src/engines/kotor2/modelloader.cpp \
     src/engines/kotor2/game.cpp \
-    src/engines/kotor2/waypoint.cpp \
-    src/engines/kotor2/situated.cpp \
-    src/engines/kotor2/placeable.cpp \
-    src/engines/kotor2/door.cpp \
     src/engines/kotor2/creature.cpp \
     src/engines/kotor2/area.cpp \
     src/engines/kotor2/objectcontainer.cpp \
     src/engines/kotor2/module.cpp \
-    src/engines/kotor2/trigger.cpp \
-    src/engines/kotor2/pathfinding.cpp \
-    src/engines/kotor2/objectwalkmesh.cpp \
-    src/engines/kotor2/doorwalkmesh.cpp \
     $(EMPTY)
 
 include src/engines/kotor2/script/rules.mk
